requirements.txt
psycopg==3.2.12
python-dotenv==1.2.1
typing_extensions==4.15.0
